+++ GDFFoundation/Models/__NEW_API/PlayerData/GDFPlayerDataSync.cs
@@ -16,7 +16,7 @@ namespace GDFFoundation
         public long Account { get; set; }
         [GDFDbAccess(updateAccess = GDFDbColumnAccess.Deny)]
         public long Reference { get; set; }
-        public short Range { get; set; }
+        public int Range { get; set; }
 
         public object GetReference() => Reference;
         public long RowId { get; set; }
